poc-low-level-overlay/python-overlay/requirements.txt
pynvim>=0.5.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
Pillow>=10.0
